+++ packages/lib/src/types/transformations.ts
@@ -1,4 +1,4 @@
-import { TransformationCase } from './transformation-case'
+import type { TransformationCase } from './transformation-case.js'
 
 export interface Transformations {
     [type: string]: {